Move git endpoints to directory
