+++ .gitignore
@@ -4,7 +4,7 @@
 # will have compiled files and executables
 debug/
 target/
-
+build/
 # Remove Cargo.lock from gitignore if creating an executable, leave it for libraries
 # More information here https://doc.rust-lang.org/cargo/guide/cargo-toml-vs-cargo-lock.html
 Cargo.lock
