+++ src/endpoints/git2/list_local_repos.rs
@@ -1,0 +1,50 @@
+use rocket::{get, State};
+use rocket::http::{ContentType, Status};
+use rocket::response::status;
+use crate::structs::AppSettings;
+
+/// *`GET /list-local-repos`*
+///
+/// Typically mounted as **`/git/list-local-repos`**
+///
+/// Returns a JSON array of local repo paths.
+///
+/// `["git.door43.org/BurritoTruck/fr_psle"]`
+#[get("/list-local-repos")]
+pub fn list_local_repos(state: &State<AppSettings>) -> status::Custom<(ContentType, String)> {
+    let root_path = state.repo_dir.lock().unwrap().clone();
+    let server_paths = std::fs::read_dir(root_path).unwrap();
+    let mut repos: Vec<String> = Vec::new();
+    for server_path in server_paths {
+        let uw_server_path_ob = server_path.unwrap().path();
+        let uw_server_path_ob2 = uw_server_path_ob.clone();
+        let server_leaf = uw_server_path_ob2.file_name().unwrap();
+        for org_path in std::fs::read_dir(uw_server_path_ob).unwrap() {
+            let uw_org_path_ob = org_path.unwrap().path();
+            let uw_org_path_ob2 = uw_org_path_ob.clone();
+            let org_leaf = uw_org_path_ob2.file_name().unwrap();
+            for repo_path in std::fs::read_dir(uw_org_path_ob).unwrap() {
+                let uw_repo_path_ob = repo_path.unwrap().path();
+                let repo_leaf = uw_repo_path_ob.file_name().unwrap();
+                let repo_url_string = format!(
+                    "{}/{}/{}",
+                    server_leaf.to_str().unwrap(),
+                    org_leaf.to_str().unwrap(),
+                    repo_leaf.to_str().unwrap()
+                );
+                repos.push(repo_url_string);
+            }
+        }
+    }
+    let quoted_repos: Vec<String> = repos
+        .into_iter()
+        .map(|str: String| format!("{}", str))
+        .collect();
+    status::Custom(
+        Status::Ok,
+        (
+            ContentType::JSON,
+            serde_json::to_string(&quoted_repos).unwrap(),
+        ),
+    )
+}
